pyproject.toml
[project]
name = "ariel"
version = "0.0.0"
description = "Ariel"
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = { text = "GPL-3.0" }
authors = [{ name = "jmdm", email = "jmdm.880@gmail.com" }]
classifiers = ["Development Status :: 2 - Pre-Alpha"]

dependencies = [
  "click>=8.0.1",
  "cmaes>=0.12.0",
  "ipykernel>=6.30.1",
  "ipywidgets>=8.1.7",
  "jupyter>=1.1.1",
  "jupyter-sphinx>=0.5.3",
  "matplotlib>=3.9.4",
  "mujoco>=3.3.5",
  "myst-nb>=1.3.0",
  "myst-parser>=3.0.1",
  "nbconvert>=7.16.6",
  "nbsphinx>=0.9.7",
  "networkx>=3.2.1",
  "nevergrad>=1.0.12",
  "nicegui>=2.23.3",
  "noise>=1.2.2",
  "nox",
  "numpy>=2.0.2",
  "numpy-quaternion>=2023.0.3",
  "numpydoc>=1.9.0",
  "opencv-python>=4.11.0.86",
  "pandoc>=2.4",
  "pillow>=11.3.0",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.10.1",
  "rich>=14.1.0",
  "shibuya>=2025.8.16",
  "sphinx-autoapi>=3.6.0",
  "sphinx-click>=6.0.0",
  "sphinx-copybutton>=0.5.2",
  "sphinx-design>=0.6.1",
  "sqlalchemy>=2.0.43",
  "sqlmodel>=0.0.24",
]


[project.urls]
Homepage = "https://github.com/ci-group/ariel"
Repository = "https://github.com/ci-group/ariel"
Documentation = "https://ariel.readthedocs.io"
Changelog = "https://github.com/ci-group/ariel/releases"

[dependency-groups]
dev = [
  "coverage[toml] >= 6.2",
  "pre-commit >=2.16.0",
  "pre-commit-hooks >=4.6.0",
  "pytest >=6.2.5",
  "pygments >=2.10.0",
]
lint = ["ruff >=0.5.5", "pydoclint >=0.5.0"]
docs = [
  "myst-parser == 3.0.1",
  "shibuya>=2025.8.16",
  "sphinx >= 4.3.2",
  "sphinx-autobuild >=2021.3.14",
  "sphinx-click >=3.0.2",
]
mypy = ["mypy >=0.930"]
typeguard = ["typeguard >=2.13.3"]
xdoctest = ["xdoctest[colors] >=0.15.10"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[project.scripts]
ariel = "ariel.__main__:cli"

[build-system]
# When using mypyc, you must list all dependencies used by the compiled
# modules here. Otherwise, mypyc will not be able to see the types of the
# dependencies, and the compilation will fail.
requires = ["setuptools>=61.0", "mypy>=1.0"]
build-backend = "setuptools.build_meta"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["ariel", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[tool.ruff]
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "B",   # flake8-bugbear
  "C90", # mccabe
  "D",   # pydocstyle
  "E",   # pycodestyle
  "F",   # pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "RUF", # Ruff-specific rules
  "S",   # flake8-bandit
  "UP",  # pyupgrade
  "W",   # pycodestyle
]
ignore = [
  "COM812",
  "COM819",
  "D107",   # undocumented-public-init conflicts with DOC301
  "D206",
  "D300",
  "E111",
  "E114",
  "E117",
  "ISC001",
  "ISC002",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "W191",
]

[tool.ruff.lint.per-file-ignores]
"*/test_*.py" = ["S101"]
"noxfile.py" = ["S101"]
"**/conftest.py" = ["S101"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pydoclint]
style = 'numpy'
exclude = '\.git|\.nox|noxfile.py'
arg-type-hints-in-docstring = false
